cloudwepapp/requirements.txt
Flask
#web server
gunicorn

mysql-connector-python
